requirement.txt
bcrypt==3.1.6
cffi==1.12.3
Click==7.0
coverage==4.5.3
entrypoints==0.3
flake8==3.7.7
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Ext==0.1
Flask-Login==0.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
peewee==3.9.5
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
python-slugify==3.0.2
six==1.12.0
text-unidecode==1.2
Werkzeug==0.15.2
WTForms==2.2.1
wtforms-html5==0.2.2
